refactor: Move code for approving queries & explanations to ApprovalTestHelpers.ts
